+++ jwst/pipeline/calwebb_coron3.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 import os.path as op
-
+from collections import defaultdict
 from ..stpipe import Pipeline
 from .. import datamodels
 from ..model_blender import blendmeta
@@ -46,19 +46,33 @@ class Coron3Pipeline(Pipeline):
 
     prefetch_references = False
 
+    # Main processing
     def process(self, user_input):
-        """Primary method for performing pipeline."""
+        """Primary method for performing pipeline.
+
+        Parameters
+        ----------
+        user_input : str, Level3 Association, or ~jwst.datamodels.DataModel
+            The exposure or association of exposures to process
+        """
         self.log.info('Starting calwebb_coron3 ...')
+        asn_exptypes = ['science', 'psf']
 
-        # Load the input association table
-        asn = self.load_as_level3_asn(user_input)
-        acid = asn.get('asn_id', '')
+        # Create a DM object using the association table
+        input_models = datamodels.open(user_input, asn_exptypes=asn_exptypes)
+        acid = input_models.meta.asn_table.asn_id
 
-        # We assume there's one final product defined by the association
-        prod = asn['products'][0]
-        self.output_file = prod.get('name', self.output_file)
+        # Store the output file for future use
+        self.output_file = input_models.meta.asn_table.products[0].name
 
-        # Setup required output products and formats
+        # Find all the member types in the product
+        members_by_type = defaultdict(list)
+        prod = input_models.meta.asn_table.products[0].instance
+
+        for member in prod['members']:
+            members_by_type[member['exptype'].lower()].append(member['expname'])
+
+        # Set up required output products and formats
         self.outlier_detection.suffix = f'{acid}_crfints'
         self.outlier_detection.save_results = self.save_results
         self.resample.blendheaders = False
@@ -68,11 +82,9 @@ class Coron3Pipeline(Pipeline):
         # processing individual inputs
         skip_outlier_detection = self.outlier_detection.skip
 
-        # Construct lists of all the PSF and science target members
-        psf_files = [m['expname'] for m in prod['members']
-                     if m['exptype'].upper() == 'PSF']
-        targ_files = [m['expname'] for m in prod['members']
-                      if m['exptype'].upper() == 'SCIENCE']
+        # Extract lists of all the PSF and science target members
+        psf_files = members_by_type['psf']
+        targ_files = members_by_type['science']
 
         # Make sure we found some PSF and target members
         if len(psf_files) == 0:
@@ -95,6 +107,7 @@ class Coron3Pipeline(Pipeline):
         for i in range(len(psf_files)):
             psf_input = datamodels.CubeModel(psf_files[i])
             psf_models.append(psf_input)
+
             psf_input.close()
 
         # Perform outlier detection on the PSFs.
@@ -114,13 +127,13 @@ class Coron3Pipeline(Pipeline):
         # Save the resulting PSF stack
         self.save_model(psf_stack, suffix='psfstack')
 
-        # Call the sequence of steps outlier_detection, align_refs, and klip
+        # Call the sequence of steps: outlier_detection, align_refs, and klip
         # once for each input target exposure
         resample_input = datamodels.ModelContainer()
         for target_file in targ_files:
             with datamodels.open(target_file) as target:
 
-                # Remove outliers from the target.
+                # Remove outliers from the target
                 if not skip_outlier_detection:
                     target = self.outlier_detection(target)
                     # step may have been skipped for this model;
@@ -158,17 +171,20 @@ class Coron3Pipeline(Pipeline):
         try:
             completed = result.meta.cal_step.resample
         except AttributeError:
-            self.log.debug('Could not determine whether resample was completed. Presuming not.')
+            self.log.debug('Could not determine if resample was completed.')
+            self.log.debug('Presuming not.')
+
             completed = 'SKIPPED'
         if completed == 'COMPLETE':
             self.log.debug(f'Blending metadata for {result}')
             blendmeta.blendmodels(result, inputs=targ_files)
 
         try:
-            result.meta.asn.pool_name = asn['asn_pool']
+            result.meta.asn.pool_name = input_models.meta.asn_table.asn_pool
             result.meta.asn.table_name = op.basename(user_input)
         except AttributeError:
-            self.log.debug(f'Cannot set association information on final result {result}')
+            self.log.debug('Cannot set association information on final')
+            self.log.debug(f'result {result}')
 
         # Save the final result
         self.save_model(result, suffix=self.suffix)